+++ test/main.test.js
@@ -9,7 +9,6 @@ var config = {
   QueueUrl: 'https://fake.us-east-1/sqs/url',
   TaskEventQueueUrl: 'https://fake.us-east-1/sqs/url-for-events',
   StackName: 'watchbot-testing',
-  ExponentialBackoff: 'false',
   AlarmOnEachFailure: 'true'
   // , LogLevel: 'debug'
 };
@@ -100,7 +99,7 @@ util.mock('[main] task running error', function(assert) {
       }
     ], 'sent expected error notification');
     util.collectionsEqual(assert, context.sqs.changeMessageVisibility, [
-      { ReceiptHandle: '1', VisibilityTimeout: 0 }
+      { ReceiptHandle: '1', VisibilityTimeout: 2 }
     ], 'expected sqs.changeMessageVisibility requests');
     assert.end();
   });
@@ -160,7 +159,7 @@ util.mock('[main] message completion error after task run failure', function(ass
     }), 'printed error message');
     util.collectionsEqual(assert, context.sns.publish, [], 'sent no error notifications');
     util.collectionsEqual(assert, context.sqs.changeMessageVisibility, [
-      { ReceiptHandle: 'error', VisibilityTimeout: 0 }
+      { ReceiptHandle: 'error', VisibilityTimeout: 2 }
     ], 'expected sqs.changeMessageVisibility requests');
     assert.end();
   });
@@ -212,7 +211,6 @@ util.mock('[main] manage messages for completed tasks', function(assert) {
 
   context.sqs.messages = [
     { MessageId: 'finish-0', ReceiptHandle: '0', Body: JSON.stringify({ Subject: 'subject0', Message: 'message0' }), Attributes: { SentTimestamp: 10, ApproximateReceiveCount: 0, ApproximateFirstReceiveTimestamp: 20 } },
-    { MessageId: 'finish-1', ReceiptHandle: '1', Body: JSON.stringify({ Subject: 'subject1', Message: 'message1' }), Attributes: { SentTimestamp: 10, ApproximateReceiveCount: 0, ApproximateFirstReceiveTimestamp: 20 } },
     { MessageId: 'finish-2', ReceiptHandle: '2', Body: JSON.stringify({ Subject: 'subject2', Message: 'message2' }), Attributes: { SentTimestamp: 10, ApproximateReceiveCount: 2, ApproximateFirstReceiveTimestamp: 20 } },
     { MessageId: 'finish-3', ReceiptHandle: '3', Body: JSON.stringify({ Subject: 'subject3', Message: 'message3' }), Attributes: { SentTimestamp: 10, ApproximateReceiveCount: 0, ApproximateFirstReceiveTimestamp: 20 } },
     { MessageId: 'finish-4', ReceiptHandle: '4', Body: JSON.stringify({ Subject: 'subject4', Message: 'message4' }), Attributes: { SentTimestamp: 10, ApproximateReceiveCount: 0, ApproximateFirstReceiveTimestamp: 20 } }
@@ -263,10 +261,9 @@ util.mock('[main] manage messages for completed tasks', function(assert) {
   setTimeout(watchbot.main.end, 1800);
   watchbot.main(testConfig).on('finish', function() {
     assert.equal(context.sqs.receiveMessage.length, 2, 'two sqs.receiveMessage requests');
-    assert.equal(context.ecs.runTask.length, 5, 'five ecs.runTask requests');
+    assert.equal(context.ecs.runTask.length, 4, 'four ecs.runTask requests');
     util.collectionsEqual(assert, context.sqs.deleteMessage, [
       { ReceiptHandle: '0-event' },
-      { ReceiptHandle: '1-event' },
       { ReceiptHandle: '2-event' },
       { ReceiptHandle: '3-event' },
       { ReceiptHandle: '4-event' },
@@ -274,8 +271,7 @@ util.mock('[main] manage messages for completed tasks', function(assert) {
       { ReceiptHandle: '3' }
     ], ' sqs.deleteMessage for all event messages, and for expected job messages');
     util.collectionsEqual(assert, context.sqs.changeMessageVisibility, [
-      { ReceiptHandle: '1', VisibilityTimeout: 0 },
-      { ReceiptHandle: '2', VisibilityTimeout: 0 },
+      { ReceiptHandle: '2', VisibilityTimeout: 8 },
       { ReceiptHandle: '4', VisibilityTimeout: 0 }
     ], 'expected sqs.changeMessageVisibility requests');
     util.collectionsEqual(assert, context.sns.publish, [
@@ -287,7 +283,7 @@ util.mock('[main] manage messages for completed tasks', function(assert) {
         Subject: config.StackName + ' failed processing message finish-3',
         Message: 'At ${date}, processing message finish-3 failed on ' + config.StackName + '\n\nTask outcome: delete & notify\n\nTask stopped reason: 3\n\nMessage information:\nMessageId: finish-3\nSubject: subject3\nMessage: message3\nSentTimestamp: 10\nApproximateFirstReceiveTimestamp: 20\nApproximateReceiveCount: 1\n\nRuntime resources:\nCluster ARN: cluster-arn\nInstance ARN: instance-arn\nTask ARN: 496a1bbc7db7ef69c5b024bed0fa66e7\n'
       }
-    ], 'expected sns.publish requests & no notification prior to NotifyAfterRetries');
+    ], 'expected sns.publish requests');
 
     assert.end();
   });