+++ test/template.test.js
@@ -32,6 +32,8 @@ test('[template] bare-bones, all defaults, no references', function(assert) {
   assert.ok(watch.Resources.WatchbotLogGroup, 'log group');
   assert.notOk(watch.Resources.WatchbotLogForwarding, 'log forwarding function');
   assert.ok(watch.Resources.WatchbotQueue, 'queue');
+  assert.ok(watch.Resources.WatchbotDeadLetterQueue, 'dead letter queue');
+  assert.ok(watch.Resources.WatchbotDeadLetterAlarm, 'dead letter alarm');
   assert.ok(watch.Resources.WatchbotTopic, 'topic');
   assert.ok(watch.Resources.WatchbotQueuePolicy, 'queue policy');
   assert.ok(watch.Resources.WatchbotQueueSizeAlarm, 'queue alarm');
@@ -46,7 +48,6 @@ test('[template] bare-bones, all defaults, no references', function(assert) {
   assert.ok(watch.Resources.WatchbotWorkerErrorsAlarm, 'worker errors alarm');
   assert.equal(watch.Resources.WatchbotWorkerErrorsAlarm.Properties.Threshold, 10, 'worker errors alarm threshold');
   assert.ok(watch.Resources.WatchbotWatcher.Properties.ContainerDefinitions[0].Environment.slice(-3, -2), 'notify after retry');
-  assert.deepEqual(watch.Resources.WatchbotWatcher.Properties.ContainerDefinitions[0].Environment.slice(-4, -3), [{ Name: 'NotifyAfterRetries', Value: 0 }], 'notify after retry default value');
   assert.notOk(watch.Resources.WatchbotWorker.Properties.ContainerDefinitions[0].Privileged, 'privileged is false');
   assert.equal(watch.Resources.WatchbotWorker.Properties.ContainerDefinitions[0].Memory, 64, 'sets default hard memory limit');
   assert.ok(watch.Resources.WatchbotWorkerRole, 'worker role');
@@ -98,7 +99,6 @@ test('[template] webhooks but no key, no references', function(assert) {
     env: { SomeKey: 'SomeValue', AnotherKey: 'AnotherValue' },
     watchers: 2,
     workers: 2,
-    backoff: false,
     mounts: '/var/tmp:/var/tmp,/mnt/data:/mnt/data,/mnt/tmp',
     reservation: {
       memory: 512,
@@ -108,7 +108,6 @@ test('[template] webhooks but no key, no references', function(assert) {
     messageRetention: 3000,
     alarmThreshold: 10,
     alarmPeriods: 6,
-    notifyAfterRetries: 2,
     privileged: true
   });
 
@@ -126,6 +125,8 @@ test('[template] webhooks but no key, no references', function(assert) {
   assert.ok(watch.Resources.testLogGroup, 'log group');
   assert.notOk(watch.Resources.testLogForwarding, 'log forwarding function');
   assert.ok(watch.Resources.testQueue, 'queue');
+  assert.ok(watch.Resources.testDeadLetterQueue, 'dead letter queue');
+  assert.ok(watch.Resources.testDeadLetterAlarm, 'dead letter alarm');
   assert.ok(watch.Resources.testTopic, 'topic');
   assert.ok(watch.Resources.testQueuePolicy, 'queue policy');
   assert.ok(watch.Resources.testQueueSizeAlarm, 'queue alarm');
@@ -140,7 +141,6 @@ test('[template] webhooks but no key, no references', function(assert) {
   assert.ok(watch.Resources.testWorkerErrorsAlarm, 'worker errors alarm');
   assert.equal(watch.Resources.testWorkerErrorsAlarm.Properties.Threshold, 10, 'worker errors alarm threshold');
   assert.ok(watch.Resources.testWatcher.Properties.ContainerDefinitions[0].Environment.slice(-3, -2), 'notify after retry');
-  assert.deepEqual(watch.Resources.testWatcher.Properties.ContainerDefinitions[0].Environment.slice(-4, -3), [{ Name: 'NotifyAfterRetries', Value: 2 }], 'notify after retry default value');
   assert.ok(watch.Resources.testWorker.Properties.ContainerDefinitions[0].Privileged, 'privileged is true');
   assert.ok(watch.Resources.testWorkerRole, 'worker role');
   assert.equal(watch.Resources.testWorkerRole.Properties.Policies.length, 1, 'default worker permissions');
@@ -190,7 +190,6 @@ test('[template] include all resources, no references', function(assert) {
     permissions: [{ Effect: 'Allow', Actions: '*', Resources: '*' }],
     watchers: 2,
     workers: 2,
-    backoff: false,
     mounts: {
       container: ['/var/tmp', '/mnt/data', '/mnt/tmp'],
       host: ['/var/tmp', '/mnt/data', '']
@@ -224,6 +223,8 @@ test('[template] include all resources, no references', function(assert) {
   assert.ok(watch.Resources.testLogGroup, 'log group');
   assert.ok(watch.Resources.testLogForwarding, 'log forwarding function');
   assert.ok(watch.Resources.testQueue, 'queue');
+  assert.ok(watch.Resources.testDeadLetterQueue, 'dead letter queue');
+  assert.ok(watch.Resources.testDeadLetterAlarm, 'dead letter alarm');
   assert.ok(watch.Resources.testTopic, 'topic');
   assert.ok(watch.Resources.testQueuePolicy, 'queue policy');
   assert.ok(watch.Resources.testQueueSizeAlarm, 'queue alarm');
@@ -295,7 +296,6 @@ test('[template] include all resources, all references', function(assert) {
     permissions: [{ Effect: 'Allow', Actions: '*', Resources: '*' }],
     watchers: cf.ref('NumWatchers'),
     workers: cf.ref('NumWorkers'),
-    backoff: cf.ref('UseBackoff'),
     mounts: {
       container: [cf.sub('/var/tmp/${stack}', { stack: cf.ref(stackName) }), '/mnt/data', '/mnt/tmp'],
       host: [cf.sub('/var/tmp/${stack}', { stack: cf.ref(stackName) }), '/mnt/data', '']
@@ -329,6 +329,8 @@ test('[template] include all resources, all references', function(assert) {
   assert.equal(watch.Resources.testLogForwarding.Condition, 'testUseLogForwarding', 'log forwarding function is conditional');
   assert.deepEqual(watch.Conditions.testUseLogForwarding, cf.notEquals(cf.ref('LogAggregationFunction'), ''), 'log forwarding condition provided');
   assert.ok(watch.Resources.testQueue, 'queue');
+  assert.ok(watch.Resources.testDeadLetterQueue, 'dead letter queue');
+  assert.ok(watch.Resources.testDeadLetterAlarm, 'dead letter alarm');
   assert.ok(watch.Resources.testTopic, 'topic');
   assert.ok(watch.Resources.testQueuePolicy, 'queue policy');
   assert.ok(watch.Resources.testQueueSizeAlarm, 'queue alarm');
@@ -355,7 +357,6 @@ test('[template] include all resources, all references', function(assert) {
   assert.ok(watch.Resources.testProgressTablePermission, 'progress table permission');
   assert.deepEqual(watch.Resources.testWorker.Properties.ContainerDefinitions[0].Environment.slice(-1), [{ Name: 'ProgressTable', Value: cf.join(['arn:aws:dynamodb:', cf.region, ':', cf.accountId, ':table/', cf.ref('testProgressTable')]) }], 'progress table env var');
   assert.deepEqual(watch.Resources.testWatcher.Properties.ContainerDefinitions[0].Environment[3], { Name: 'Concurrency', Value: cf.ref('NumWorkers') }, 'sets Concurrency');
-  assert.deepEqual(watch.Resources.testWatcher.Properties.ContainerDefinitions[0].Environment[8], { Name: 'ExponentialBackoff', Value: cf.ref('UseBackoff') }, 'sets ExponentialBackoff');
   assert.deepEqual(watch.Resources.testWatcher.Properties.ContainerDefinitions[0].Environment.slice(-1), [{ Name: 'AlarmOnEachFailure', Value: cf.ref('AlarmOnFailures') }], 'alarm on failure env var');
   assert.deepEqual(watch.Resources.testWorker.Properties.Volumes[0], { Name: 'mnt-0', Host: { SourcePath: { 'Fn::Sub': ['/var/tmp/${stack}', { stack: { Ref: 'some-stack-name' } }] } } });
   assert.deepEqual(watch.Resources.testWorker.Properties.Volumes[1], { Name: 'mnt-1', Host: { SourcePath: '/mnt/data' } });
@@ -393,7 +394,6 @@ test('[template] resources are valid', function(assert) {
     permissions: [{ Effect: 'Allow', Actions: '*', Resources: '*' }],
     watchers: 2,
     workers: 2,
-    backoff: false,
     mounts: '/var/tmp:/var/tmp,/mnt/data:/mnt/data,/mnt/tmp',
     logAggregationFunction: 'arn:aws:lambda:us-east-1:123456789000:function:log-fake-test',
     reservation: {
@@ -437,7 +437,6 @@ test('[template] notificationTopic vs notificationEmail', function(assert) {
       permissions: [{ Effect: 'Allow', Actions: '*', Resources: '*' }],
       watchers: 2,
       workers: 2,
-      backoff: false,
       mounts: '/var/tmp:/var/tmp,/mnt/data:/mnt/data,/mnt/tmp',
       logAggregationFunction: 'arn:aws:lambda:us-east-1:123456789000:function:log-fake-test',
       reservation: {
@@ -467,7 +466,6 @@ test('[template] notificationTopic vs notificationEmail', function(assert) {
     permissions: [{ Effect: 'Allow', Actions: '*', Resources: '*' }],
     watchers: 2,
     workers: 2,
-    backoff: false,
     mounts: '/var/tmp:/var/tmp,/mnt/data:/mnt/data,/mnt/tmp',
     logAggregationFunction: 'arn:aws:lambda:us-east-1:123456789000:function:log-fake-test',
     reservation: {