+++ changelog.md
@@ -13,6 +13,9 @@
 - **BREAKING** by default, watchbot no longer sends notification emails each time a worker errors. You can opt-in to this behavior by setting `watchbot.template(options)` `.alarmOnEachFailure: true`.
 - **BREAKING** no longer sends notifications on error interacting with SQS. Instead watchbot silently proceeds.
 - **BREAKING** watcher log format has changed. Now watcher logs print JSON objects
+- **BREAKING** removes `.notifyAfterRetries` option
+- **BREAKING** removes `.backoff` option. Workers are always retried with exponential backoff
+- **BREAKING** adds a dead letter queue. Messages received more than 14 times by a watcher container will be sent to this queue. Any visible messages in this queue will trip an alarm.
 
 ### 1.4.0